bugfix coord transf dtype

+++ source/data_transforms/_frad_transforms.py
@@ -132,6 +132,6 @@ def frad(data:AtomicData, add_coords_noise: bool = True):
 
 def coord_noise(data:AtomicData):
     data = deepcopy(data) #! in-place op to the data obj persist thru dloader iterations
-    data.noise_target = torch.from_numpy(np.random.normal(0, 1, size=data.pos.shape)) * 0.005
+    data.noise_target = torch.from_numpy(np.random.normal(0, 1, size=data.pos.shape) * 0.005).to(torch.float32)
     data.pos += data.noise_target
     return data
